feat: Change handling page, dimen variables
