+++ app/src/main/java/com/doublejj/edit/ui/modules/main/home/open_comment/OpenCommentActivity.kt
@@ -32,7 +32,7 @@ class OpenCommentActivity : AppCompatActivity(), CommentsOfSentenceView {
 
         /** toolbar buttons **/
         binding.ibBack.setOnClickListener {
-            supportFragmentManager.popBackStack()
+            finish()
         }
         binding.ibRefresh.setOnClickListener {
             // refresh data
