+++ app/src/main/res/layout/activity_enter_code.xml
@@ -68,7 +68,7 @@
                 android:gravity="center"
                 android:text="@string/tv_info_enter_code"
                 android:textColor="@color/black"
-                android:textSize="@dimen/tv_18sp" />
+                android:textSize="@dimen/size_18dp" />
         </LinearLayout>
 
         <!--Spinner-->
@@ -86,7 +86,7 @@
                 android:fontFamily="@font/noto_sans_kr_regular"
                 android:text="@string/tv_code"
                 android:textColor="@color/black"
-                android:textSize="@dimen/tv_15sp" />
+                android:textSize="@dimen/size_15dp" />
 
             <LinearLayout
                 android:layout_width="match_parent"
@@ -106,7 +106,7 @@
                     android:inputType="textPersonName|textNoSuggestions"
                     android:paddingHorizontal="@dimen/hint_padding_10dp"
                     android:paddingVertical="@dimen/padding_2.5dp"
-                    android:textSize="@dimen/tv_15sp" />
+                    android:textSize="@dimen/size_15dp" />
 
                 <TextView
                     android:id="@+id/tv_counter_enter_code"
@@ -126,7 +126,7 @@
                 android:fontFamily="@font/noto_sans_kr_regular"
                 android:text="@string/tv_caption_enter_code"
                 android:textColor="@color/very_light_pink"
-                android:textSize="@dimen/tv_10sp" />
+                android:textSize="@dimen/size_10dp" />
 
         </LinearLayout>
