+++ app/src/main/res/layout/activity_info_first.xml
@@ -53,7 +53,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_name_info"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <EditText
                             android:id="@+id/et_name_info_first"
@@ -67,7 +67,7 @@
                             android:inputType="textPersonName|textNoSuggestions"
                             android:paddingHorizontal="@dimen/hint_padding_10dp"
                             android:paddingVertical="@dimen/padding_2.5dp"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <TextView
                             android:id="@+id/tv_name_caption_info_first"
@@ -76,7 +76,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_name_caption_info"
                             android:textColor="@color/very_light_pink"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
                     </LinearLayout>
 
                     <!-- Nick_Name -->
@@ -93,7 +93,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_nick_name_info"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <!-- Double Check Nick Name -->
                         <LinearLayout
@@ -117,7 +117,7 @@
                                 android:inputType="textPersonName|textNoSuggestions"
                                 android:paddingHorizontal="@dimen/hint_padding_10dp"
                                 android:paddingVertical="@dimen/padding_2.5dp"
-                                android:textSize="@dimen/tv_15sp" />
+                                android:textSize="@dimen/size_15dp" />
 
                             <androidx.appcompat.widget.AppCompatButton
                                 android:id="@+id/btn_double_check"
@@ -138,7 +138,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_nick_name_caption_info"
                             android:textColor="@color/very_light_pink"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
                     </LinearLayout>
 
                     <!-- Phone -->
@@ -157,7 +157,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_phone_info"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <EditText
                             android:id="@+id/et_phone_info_first"
@@ -171,7 +171,7 @@
                             android:inputType="phone|textNoSuggestions"
                             android:paddingHorizontal="@dimen/hint_padding_10dp"
                             android:paddingVertical="@dimen/padding_2.5dp"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <TextView
                             android:id="@+id/tv_phone_caption_info_first"
@@ -180,7 +180,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_phone_caption_info"
                             android:textColor="@color/very_light_pink"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
                     </LinearLayout>
 
 
@@ -205,7 +205,7 @@
                         android:paddingHorizontal="@dimen/padding_20dp"
                         android:text="@string/tv_privacy_info_first"
                         android:textColor="@color/black"
-                        android:textSize="@dimen/tv_12sp"
+                        android:textSize="@dimen/size_12dp"
                         tools:ignore="MissingConstraints" />
 
                     <!-- Button -->
