+++ app/src/main/res/layout/activity_info_second.xml
@@ -52,7 +52,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_email_info"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <!-- Double Check Nick Name -->
                         <LinearLayout
@@ -76,7 +76,7 @@
                                 android:inputType="textPersonName"
                                 android:paddingHorizontal="@dimen/hint_padding_10dp"
                                 android:paddingVertical="@dimen/padding_2.5dp"
-                                android:textSize="@dimen/tv_15sp" />
+                                android:textSize="@dimen/size_15dp" />
 
                             <androidx.appcompat.widget.AppCompatButton
                                 android:id="@+id/btn_double_check"
@@ -97,7 +97,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_email_caption_info"
                             android:textColor="@color/very_light_pink"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
                     </LinearLayout>
 
                     <!-- PW -->
@@ -116,7 +116,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_pw_info"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <EditText
                             android:id="@+id/et_pw_info_second"
@@ -130,7 +130,7 @@
                             android:inputType="textPassword"
                             android:paddingHorizontal="@dimen/hint_padding_10dp"
                             android:paddingVertical="@dimen/padding_2.5dp"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <TextView
                             android:id="@+id/tv_pw_caption_info_second"
@@ -139,7 +139,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_pw_caption_info"
                             android:textColor="@color/very_light_pink"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
                     </LinearLayout>
 
                     <!-- re PW -->
@@ -158,7 +158,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_re_pw_info"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <EditText
                             android:id="@+id/et_re_pw_info_first"
@@ -172,7 +172,7 @@
                             android:inputType="textPassword"
                             android:paddingHorizontal="@dimen/hint_padding_10dp"
                             android:paddingVertical="@dimen/padding_2.5dp"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <TextView
                             android:id="@+id/tv_re_pw_caption_info_first"
@@ -181,7 +181,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/hint_re_pw_info"
                             android:textColor="@color/very_light_pink"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
                     </LinearLayout>
