+++ app/src/main/res/layout/activity_log_in.xml
@@ -58,7 +58,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_email_info"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
 
                         <EditText
@@ -73,7 +73,7 @@
                             android:inputType="textPersonName"
                             android:paddingHorizontal="@dimen/hint_padding_10dp"
                             android:paddingTop="@dimen/hint_padding_5dp"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <TextView
                             android:id="@+id/tv_email_caption_log_in"
@@ -83,7 +83,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_email_caption_info"
                             android:textColor="@color/very_light_pink"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
                     </LinearLayout>
 
                     <!--PW-->
@@ -102,7 +102,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_pw_info"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <EditText
                             android:id="@+id/et_pw_log_in"
@@ -115,7 +115,7 @@
                             android:inputType="textPassword"
                             android:paddingHorizontal="@dimen/hint_padding_10dp"
                             android:paddingTop="@dimen/hint_padding_5dp"
-                            android:textSize="@dimen/tv_15sp" />
+                            android:textSize="@dimen/size_15dp" />
 
                         <TextView
                             android:id="@+id/tv_pw_caption_log_in"
@@ -124,7 +124,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_pw_caption_info_log_in"
                             android:textColor="@color/very_light_pink"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
                     </LinearLayout>
                 </LinearLayout>
 
@@ -154,7 +154,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_info_find_email"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
 
                         <TextView
                             android:id="@+id/tv_find_email_log_in"
@@ -164,7 +164,7 @@
                             android:onClick="onClick"
                             android:text="@string/tv_find_email"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_10sp"
+                            android:textSize="@dimen/size_10dp"
                             android:clickable="true"
                             android:textStyle="bold" />
                     </LinearLayout>
@@ -184,7 +184,7 @@
                             android:fontFamily="@font/noto_sans_kr_regular"
                             android:text="@string/tv_info_find_pw"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_10sp" />
+                            android:textSize="@dimen/size_10dp" />
 
                         <TextView
                             android:id="@+id/tv_find_pw_log_in"
@@ -195,7 +195,7 @@
                             android:clickable="true"
                             android:text="@string/tv_find_pw"
                             android:textColor="@color/black"
-                            android:textSize="@dimen/tv_10sp"
+                            android:textSize="@dimen/size_10dp"
                             android:textStyle="bold" />
                     </LinearLayout>
                 </LinearLayout>
