+++ app/src/main/res/layout/activity_my_sympathized_sentence.xml
@@ -1,0 +1,94 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <androidx.appcompat.widget.Toolbar
+                android:id="@+id/toolbar_title_with_buttons"
+                android:layout_width="match_parent"
+                android:layout_height="@dimen/size_toolbar_height"
+                app:contentInsetStart="@dimen/margin_component_inset_20dp"
+                app:contentInsetEnd="@dimen/margin_component_inset_20dp"
+                android:background="@color/white"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent">
+                <androidx.constraintlayout.widget.ConstraintLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content">
+                    <ImageButton
+                        android:id="@+id/ib_back"
+                        android:layout_width="@dimen/size_icon_button_24dp"
+                        android:layout_height="@dimen/size_icon_button_24dp"
+                        android:src="@drawable/icon_back"
+                        android:background="@color/transparent"
+                        android:scaleType="fitXY"
+
+                        app:layout_constraintStart_toStartOf="parent"
+                        app:layout_constraintTop_toTopOf="parent"
+                        app:layout_constraintBottom_toBottomOf="parent" />
+                    <TextView
+                        android:id="@+id/tv_title"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        style="@style/CustomTextViewMidBlack"
+                        android:text="@string/tv_menu_sympathy_list"
+                        android:textSize="@dimen/size_tv_title_15dp"
+                        android:fontFamily="@font/noto_sans_kr_medium"
+                        android:background="@color/transparent"
+                        android:gravity="center"
+
+                        app:layout_constraintStart_toStartOf="parent"
+                        app:layout_constraintEnd_toEndOf="parent"
+                        app:layout_constraintTop_toTopOf="parent"
+                        app:layout_constraintBottom_toBottomOf="parent" />
+                    <ImageButton
+                        android:id="@+id/ib_refresh"
+                        android:layout_width="@dimen/size_icon_button_24dp"
+                        android:layout_height="@dimen/size_icon_button_24dp"
+                        android:src="@drawable/icon_refresh"
+                        android:background="@color/transparent"
+                        android:scaleType="fitXY"
+
+                        app:layout_constraintEnd_toEndOf="parent"
+                        app:layout_constraintTop_toTopOf="parent"
+                        app:layout_constraintBottom_toBottomOf="parent" />
+                </androidx.constraintlayout.widget.ConstraintLayout>
+            </androidx.appcompat.widget.Toolbar>
+
+            <androidx.core.widget.NestedScrollView
+                android:layout_width="0dp"
+                android:layout_height="0dp"
+
+                android:background="@color/gray_background"
+
+                app:layout_constraintTop_toBottomOf="@id/toolbar_title_with_buttons"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent">
+
+                <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:orientation="vertical"
+                    android:background="@color/gray_background">
+
+                    <!--  sentence section  -->
+                    <androidx.recyclerview.widget.RecyclerView
+                        android:id="@+id/rv_sentence"
+                        android:layout_width="match_parent"
+                        android:layout_height="match_parent"
+
+                        tools:listitem="@layout/layout_sentence" />
+                </LinearLayout>
+            </androidx.core.widget.NestedScrollView>
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </FrameLayout>
+</layout>
