+++ app/src/main/res/layout/activity_privacy_agree.xml
@@ -84,7 +84,7 @@
                     android:gravity="center"
                     android:text="@string/tv_info_privacy_agree"
                     android:textColor="@color/black"
-                    android:textSize="@dimen/tv_18sp" />
+                    android:textSize="@dimen/size_18dp" />
             </LinearLayout>
         </LinearLayout>
