+++ app/src/main/res/layout/activity_pw_find.xml
@@ -71,7 +71,7 @@
                     android:gravity="center"
                     android:text="@string/tv_title_email_find"
                     android:textColor="@color/black"
-                    android:textSize="@dimen/tv_18sp" />
+                    android:textSize="@dimen/size_18dp" />
             </LinearLayout>
 
             <!--Enter Content-->
@@ -99,7 +99,7 @@
                         android:text="@string/tv_name_info"
                         android:textColor="@color/black"
                         android:paddingVertical="@dimen/padding_2.5dp"
-                        android:textSize="@dimen/tv_15sp" />
+                        android:textSize="@dimen/size_15dp" />
 
                     <EditText
                         android:id="@+id/et_name_pw_find"
@@ -113,7 +113,7 @@
                         android:inputType="textPersonName|textNoSuggestions"
                         android:paddingHorizontal="@dimen/hint_padding_10dp"
                         android:paddingVertical="2.5dp"
-                        android:textSize="@dimen/tv_15sp" />
+                        android:textSize="@dimen/size_15dp" />
 
                     <TextView
                         android:id="@+id/tv_name_caption_pw_find"
@@ -122,7 +122,7 @@
                         android:fontFamily="@font/noto_sans_kr_regular"
                         android:text="@string/tv_name_caption_info"
                         android:textColor="@color/very_light_pink"
-                        android:textSize="@dimen/tv_10sp" />
+                        android:textSize="@dimen/size_10dp" />
                 </LinearLayout>
 
                 <!--Email-->
@@ -138,7 +138,7 @@
                         android:fontFamily="@font/noto_sans_kr_regular"
                         android:text="@string/tv_email_info"
                         android:textColor="@color/black"
-                        android:textSize="@dimen/tv_15sp" />
+                        android:textSize="@dimen/size_15dp" />
 
                     <EditText
                         android:id="@+id/et_email_pw_find"
@@ -152,7 +152,7 @@
                         android:paddingHorizontal="@dimen/hint_padding_10dp"
                         android:paddingVertical="@dimen/padding_2.5dp"
                         android:hint="@string/hint_email_info"
-                        android:textSize="@dimen/tv_15sp" />
+                        android:textSize="@dimen/size_15dp" />
 
                     <TextView
                         android:id="@+id/tv_caption_email_pw_find"
@@ -161,7 +161,7 @@
                         android:fontFamily="@font/noto_sans_kr_regular"
                         android:text="@string/tv_email_was_caption_info"
                         android:textColor="@color/very_light_pink"
-                        android:textSize="@dimen/tv_10sp" />
+                        android:textSize="@dimen/size_10dp" />
 
                 </LinearLayout>
 
@@ -181,7 +181,7 @@
                         android:fontFamily="@font/noto_sans_kr_regular"
                         android:text="@string/tv_phone_info"
                         android:textColor="@color/black"
-                        android:textSize="@dimen/tv_15sp" />
+                        android:textSize="@dimen/size_15dp" />
 
                     <EditText
                         android:id="@+id/et_phone_pw_find"
@@ -195,7 +195,7 @@
                         android:inputType="phone|textNoSuggestions"
                         android:paddingHorizontal="@dimen/hint_padding_10dp"
                         android:paddingVertical="@dimen/padding_2.5dp"
-                        android:textSize="@dimen/tv_15sp" />
+                        android:textSize="@dimen/size_15dp" />
 
                     <TextView
                         android:id="@+id/tv_phone_caption_pw_find"
@@ -204,7 +204,7 @@
                         android:fontFamily="@font/noto_sans_kr_regular"
                         android:text="@string/tv_phone_caption_info"
                         android:textColor="@color/very_light_pink"
-                        android:textSize="@dimen/tv_10sp" />
+                        android:textSize="@dimen/size_10dp" />
                 </LinearLayout>
             </LinearLayout>
