+++ app/src/main/res/layout/activity_read.xml
@@ -76,7 +76,7 @@
                     android:gravity="center"
                     android:text="@string/tv_mentor_info_read"
                     android:textColor="@color/black"
-                    android:textSize="@dimen/tv_18sp" />
+                    android:textSize="@dimen/size_18dp" />
             </LinearLayout>
 
             <TextView
@@ -119,7 +119,7 @@
                         android:fontFamily="@font/noto_sans_kr_medium"
                         android:text="@string/tv_one_title_read"
                         android:textColor="@color/purple"
-                        android:textSize="@dimen/tv_13sp"
+                        android:textSize="@dimen/size_13dp"
                         android:textStyle="normal" />
 
                     <TextView
@@ -131,7 +131,7 @@
                         android:paddingLeft="@dimen/size_15dp"
                         android:text="@string/tv_one_content_read"
                         android:textColor="@color/black"
-                        android:textSize="@dimen/tv_12sp"
+                        android:textSize="@dimen/size_12dp"
                         android:textStyle="normal" />
                 </LinearLayout>
 
@@ -149,7 +149,7 @@
                         android:fontFamily="@font/noto_sans_kr_medium"
                         android:text="@string/tv_two_title_read"
                         android:textColor="@color/purple"
-                        android:textSize="@dimen/tv_13sp"
+                        android:textSize="@dimen/size_13dp"
                         android:textStyle="normal" />
 
                     <TextView
@@ -161,7 +161,7 @@
                         android:paddingLeft="@dimen/size_15dp"
                         android:text="@string/tv_two_content_read"
                         android:textColor="@color/black"
-                        android:textSize="@dimen/tv_12sp"
+                        android:textSize="@dimen/size_12dp"
                         android:textStyle="normal" />
                 </LinearLayout>
 
@@ -178,7 +178,7 @@
                         android:fontFamily="@font/noto_sans_kr_medium"
                         android:text="@string/tv_three_title_read"
                         android:textColor="@color/purple"
-                        android:textSize="@dimen/tv_13sp"
+                        android:textSize="@dimen/size_13dp"
                         android:textStyle="normal" />
 
                     <TextView
@@ -190,7 +190,7 @@
                         android:paddingLeft="@dimen/size_15dp"
                         android:text="@string/tv_three_content_read"
                         android:textColor="@color/black"
-                        android:textSize="@dimen/tv_12sp"
+                        android:textSize="@dimen/size_12dp"
                         android:textStyle="normal" />
                 </LinearLayout>
             </LinearLayout>
@@ -225,7 +225,7 @@
                     android:lineSpacingExtra="5sp"
                     android:text="@string/tv_check_info_read"
                     android:textColor="@color/black"
-                    android:textSize="@dimen/tv_12sp"
+                    android:textSize="@dimen/size_12dp"
                     android:textStyle="normal" />
 
             </LinearLayout>
